refactor internal
