+++ internal/config/config_test.go
@@ -1,0 +1,107 @@
+package config
+
+import "testing"
+
+func Test_encodeAuth(t *testing.T) {
+	tests := []struct {
+		name     string
+		username string
+		password string
+		want     string
+	}{
+		{
+			name:     "Username and password",
+			username: "username",
+			password: "password",
+			want:     "dXNlcm5hbWU6cGFzc3dvcmQ=",
+		},
+		{
+			name:     "Username only",
+			username: "username",
+			password: "",
+			want:     "dXNlcm5hbWU6",
+		},
+		{
+			name:     "Password only",
+			username: "",
+			password: "password",
+			want:     "OnBhc3N3b3Jk",
+		},
+		{
+			name:     "Empty username and empty password",
+			username: "",
+			password: "",
+			want:     "",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := encodeAuth(tt.username, tt.password); got != tt.want {
+				t.Errorf("encodeAuth() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func Test_decodeAuth(t *testing.T) {
+	tests := []struct {
+		name     string
+		authStr  string
+		username string
+		password string
+		wantErr  bool
+	}{
+		{
+			name:     "Valid base64",
+			authStr:  "dXNlcm5hbWU6cGFzc3dvcmQ=", // username:password
+			username: "username",
+			password: "password",
+		},
+		{
+			name:     "Valid base64, username only",
+			authStr:  "dXNlcm5hbWU6", // username:
+			username: "username",
+		},
+		{
+			name:     "Valid base64, password only",
+			authStr:  "OnBhc3N3b3Jk", // :password
+			password: "password",
+		},
+		{
+			name:     "Valid base64, bad format",
+			authStr:  "d2hhdGV2ZXI=", // whatever
+			username: "",
+			password: "",
+			wantErr:  true,
+		},
+		{
+			name:     "Invalid base64",
+			authStr:  "whatever",
+			username: "",
+			password: "",
+			wantErr:  true,
+		},
+		{
+			name:     "Empty string",
+			authStr:  "",
+			username: "",
+			password: "",
+			wantErr:  false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			gotUsername, gotPassword, err := decodeAuth(tt.authStr)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("decodeAuth() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if gotUsername != tt.username {
+				t.Errorf("decodeAuth() got = %v, want %v", gotUsername, tt.username)
+			}
+			if gotPassword != tt.password {
+				t.Errorf("decodeAuth() got1 = %v, want %v", gotPassword, tt.password)
+			}
+		})
+	}
+}
